old_requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.4.0
appdirs==1.4.3
appier==1.21.7
asgiref==3.4.1
astroid==2.5.6
async-timeout==4.0.2
attrs==21.2.0
# backports.zoneinfo==0.2.1
bidict==0.21.4
boto3==1.17.72
botocore==1.20.72
CacheControl==0.12.6
cachetools==4.2.2
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
Django==4.0
dnspython==1.16.0
eventlet==0.30.2
facebook-sdk==3.1.0
fastapi==0.70.1
Flask==2.0.2
Flask-SocketIO==5.1.1
frozenlist==1.2.0
google-api==0.1.12
google-api-core==1.26.3
google-api-python-client==2.4.0
google-auth==1.30.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
html5lib==1.0.1
httplib2==0.19.1
httptools==0.3.0
idna==3.3
ipaddr==2.2.0
isort==5.8.0
itsdangerous==2.0.1
Jinja2==3.0.3
jmespath==0.10.0
lazy-object-proxy==1.6.0
lockfile==0.12.2
MarkupSafe==2.0.1
mccabe==0.6.1
msgpack==0.6.2
multidict==5.2.0
netifaces==0.10.6
packaging==20.3
pep517==0.8.2
progress==1.5
protobuf==3.17.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
# pycrypto==2.6.1
pycryptodome==3.10.1
pydantic==1.8.2
pyfcm==1.5.3
PyJWT==2.1.0
pylint==2.8.2
pymongo==4.0.1
pyparsing==2.4.6
python-dateutil==2.8.1
python-dotenv==0.19.2
python-engineio==4.3.0
python-multipart==0.0.5
python-socketio==5.5.0
pytoml==0.1.21
pytz==2021.1
PyYAML==6.0
requests==2.26.0
retrying==1.3.3
rope==0.19.0
rsa==4.7.2
s3transfer==0.4.2
six==1.14.0
sniffio==1.2.0
sqlparse==0.4.2
starlette==0.16.0
toml==0.10.2
typing-extensions==4.0.1
tzdata==2021.5
uritemplate==3.0.1
urllib3==1.26.7
uvicorn==0.16.0
# uvloop==0.15.2
watchgod==0.7
webencodings==0.5.1
websocket-client==1.2.3
websockets==10.1
Werkzeug==2.0.2
# wrapt==1.12.1
yarl==1.7.2
